# code/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api2/v1/aggregation/{aggregation_type}/student/{student_id}": {
            "get": {
                "operationId": "cf1a9f854a37f769e1aeaf80e2238997",
                "parameters": [
                    {
                        "name": "student_id",
                        "in": "path",
                        "description": "The student-identifier.",
                        "required": true,
                        "example": 1
                    },
                    {
                        "name": "aggregation_type",
                        "in": "path",
                        "description": "types: time-today, time-month, skills, courses",
                        "required": true,
                        "example": "time-today"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api2/v1/student/{id}/rate-task": {
            "post": {
                "operationId": "79ce6eaafe98b099ebee82ae28cc785c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The identifier of student.",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "task": {
                                        "example": "1"
                                    },
                                    "rating": {
                                        "example": 37
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}